@@ -2,10 +2,12 @@ export type ProfileSource = 'signup' | 'script' | 'migration' | 'signup-existing
 
 export interface UserProfile {
   email: string;
-  createdAt: number;
+  createdAt: string; // ISO date YYYY-MM-DD
   source: ProfileSource;
   serverId?: number;
   displayName?: string;
+  firstName?: string;
+  lastName?: string;
 }
 
 export interface SensorReading {
@@ -14,6 +16,7 @@ export interface SensorReading {
   temperature?: number;
   humidity?: number;
   ts: number;
+  date?: string; // YYYY-MM-DD
   meta?: Record<string, any>;
 }
 
