@@ -4,7 +4,13 @@ import { AuthController } from '../controllers/auth.controller';
 const router = Router();
 const authController = new AuthController();
 
-router.post('/login', authController.loginUser);
-router.post('/logout', authController.logoutUser);
+// Signup (create new user)
+router.post('/signup', authController.signupUser);
+
+// Signin / login
+router.post('/signin', authController.loginUser);
+
+// Signout / logout
+router.post('/signout', authController.logoutUser);
 
 export default router;
